extract Left- & RightExtension into separate module
